CSS +  Polishing + refining skeleton
I further refined the html skeleton. It's not fully complete as I haven't received all the information yet. But I've added a navigation bar, and plan to create separate html pages for the different.

The different pages will be:
1. Home page
2. Practical business application
3. About the team.

------
I improved the current heading and top navigation bar. I added the logo as well, removing the previous text "Welcome to..."

The colour pallet has also changed, and will now be focused on grey, white colours.