About the team section
I created the "About the Team section".
The format for a block consisted of a heading(the persons name), an image of the person, a description of them/ their work, experience and then their linkedin at the end.
There's also a small segment describing what brings the team together.

TODO: rework the scaling of the images and the team's names text. When zoomed out at certain point, the text vanishes into the image that doesn't scale; just keeps its proportions.