Altered Feature's and What's it about pages
My client has sent the updated text and layout of the "features" and the "what's it about" page.

Two blocks/paragraphs in the "Features" were removed, while one new block was added.

Also changed CSS of H2 on each page that features it. Text-align centre was missing for some reason :0